Moved transpiler into cellmlmanip main module.
