--- cellmlmanip/mathml2sympy/__init__.py
@@ -1,5 +1,0 @@
-"""CellML uses content MathML for mathematical expressions. This module
-translates a subset of MathML (as used by Cardiac Electrophysiology Web Lab)
-to SymPy expressions.
-"""
-from .transpiler import Transpiler  # noqa
